+++ activejob/CHANGELOG.md
@@ -1,3 +1,26 @@
+*   Add structured events for Active Job:
+    - `active_job.enqueue_failed`
+    - `active_job.enqueue_aborted`
+    - `active_job.enqueued`
+    - `active_job.bulk_enqueued`
+    - `active_job.started`
+    - `active_job.failed`
+    - `active_job.aborted`
+    - `active_job.completed`
+    - `active_job.retry_scheduled`
+    - `active_job.retry_stopped`
+    - `active_job.discarded`
+    - `active_job.interrupt`
+    - `active_job.resume`
+    - `active_job.step_skipped`
+    - `active_job.step_resumed`
+    - `active_job.step_started`
+    - `active_job.step_interrupted`
+    - `active_job.step_errored`
+    - `active_job.step`
+
+    *Adrianna Chang*
+
 ## Rails 8.1.0.beta1 (September 04, 2025) ##
 
 *   Deprecate built-in `sidekiq` adapter.